requirements.txt
Flask==2.3.2
tensorflow==2.13.0
Pillow==10.0.0
numpy==1.24.3
gunicorn==20.1.0
flask-cors==3.0.10